Move _create_broadcasted_txn to client utils
